build(publish-docs.js): improve script for publishing docz-app to GitHub Pages
Add several options to the script, including a help-option, which is run with `npm run docz:publish
-- --h`.

+++ publish-docs.js
@@ -1,14 +1,77 @@
 const shell = require('shelljs');
 const fs = require('fs');
-const argv = require('yargs').argv;
-const version = argv.v || '0.0.0-dev';
-const dryRun = argv.dryRun === 'true';
 const replace = require('replace-in-file');
+const argv = require('yargs').argv;
 
-setup();
-build();
-commitAndPush();
-teardown();
+const version = argv.v || '0.0.0-dev';
+
+const pruneDev = argv.pruneDev !== undefined;
+const removeSpecific = !!argv.remove;
+
+const runSetup = argv.noSetup === undefined;
+const runBuild = argv.noBuild === undefined;
+const runCommit = argv.noCommit === undefined;
+const runPush = argv.noPush === undefined;
+const runTeardown = argv.noTeardown === undefined;
+
+const cleanOnFail = runTeardown && argv.noCleanOnFail === undefined;
+const dryRun = argv.dryRun !== undefined;
+
+if (argv.h !== undefined) {
+    shell.echo(`
+usage: npm run docz:publish [-- [--v=<version>] [--remove=<pattern>] [--pruneDev]
+                                [--noSetup] [--noBuild] [--noCommit] [--noPush]
+                                [--noTeardown] [--dryRun] [--noCleanOnFail]]
+
+    --v             The version number for this release of the documentation.
+                    Defaults to '0.0.0-dev'.
+    --remove        Removes all versions matching the given filename-pattern.
+    --pruneDev      Alias for --removeSpecific=0.0.0-dev*
+    --noSetup       Run no setup. Only use this if you have previously run the setup step
+                    without running the teardown step afterward.
+    --noBuild       Do not build the documentation.
+    --noCommit      Do not commit any changes.
+    --noPush        Do not push any commits.
+    --noTeardown    Run no cleanup at end of script. Implies --noCleanOnFail.
+    --noCleanOnFail Do not run cleanup if script fails. Unless --noTeardown is set,
+                    cleanup will still be run if script is successful.
+    `);
+} else if (removeSpecific || runPruneDev) {
+    if (runSetup) {
+        setup();
+    }
+    if (pruneDev) {
+        remove('0.0.0-dev*');
+    }
+    if (removeSpecific) {
+        remove(argv.remove);
+    }
+    if (runCommit) {
+        commit('chore(deploy docs): prune dev-versions');
+    }
+    if (runPush) {
+        push();
+    }
+    if (runTeardown) {
+        teardown(true);
+    }
+} else {
+    if (runSetup) {
+        setup();
+    }
+    if (runBuild) {
+        build();
+    }
+    if (runCommit) {
+        commit();
+    }
+    if (runPush) {
+        push();
+    }
+    if (runTeardown) {
+        teardown(true);
+    }
+}
 
 function setup() {
     if (!shell.which('git')) {
@@ -42,10 +105,12 @@ function setup() {
         teardown();
         shell.exit(1);
     }
+
+    shell.cd('..');
 }
 
 function build() {
-    shell.cd('versions');
+    shell.cd('docsDist/versions');
 
     shell.echo('Removing old version folder if it already exists.');
     shell.rm('-rf', version);
@@ -103,6 +168,17 @@ function build() {
         shell.exit(1);
     }
 
+    updateVersionList();
+}
+
+function remove(pattern) {
+    shell.cd('docsDist/versions');
+    shell.rm('-rf', pattern);
+    shell.cd('../..');
+    updateVersionList();
+}
+
+function updateVersionList() {
     shell.cd('docsDist');
 
     const files = fs.readdirSync('versions');
@@ -110,14 +186,19 @@ function build() {
         'versions.js',
         `window.versions = ${JSON.stringify(files)};`
     );
+
+    shell.cd('..');
 }
 
-function commitAndPush() {
+function commit(message) {
+    message = message || 'chore(deploy docs): deploy latest docs to gh-pages';
+    shell.cd('docsDist');
+
     shell.exec('git add -A --ignore-errors');
 
     if (
         shell.exec(
-            `git commit --author "$GIT_AUTHOR_NAME <$GIT_AUTHOR_EMAIL>" -m "chore(deploy docs): deploy latest docs to gh-pages"`
+            `git commit --author "$GIT_AUTHOR_NAME <$GIT_AUTHOR_EMAIL>" -m "${message}"`
         ).code !== 0
     ) {
         shell.echo('git commit failed!');
@@ -125,6 +206,12 @@ function commitAndPush() {
         teardown();
         shell.exit(1);
     }
+
+    shell.cd('..');
+}
+
+function push() {
+    shell.cd('docsDist');
 
     if (dryRun) {
         shell.exec('git log -1');
@@ -143,12 +230,14 @@ function commitAndPush() {
     shell.cd('..');
 }
 
-function teardown() {
-    shell.exec(
-        'git checkout doczrc.js src/examples/example.tsx src/examples/props.tsx src/index.html stencil.config.docs.ts'
-    );
-    shell.echo('Removing worktree for docsDist.');
-    shell.exec('git worktree remove docsDist --force');
-    shell.echo('Deleting local branch gh-pages.');
-    shell.exec('git branch -D gh-pages');
+function teardown(finished) {
+    if (finished || cleanOnFail) {
+        shell.exec(
+            'git checkout doczrc.js src/examples/example.tsx src/examples/props.tsx src/index.html stencil.config.docs.ts'
+        );
+        shell.echo('Removing worktree for docsDist.');
+        shell.exec('git worktree remove docsDist --force');
+        shell.echo('Deleting local branch gh-pages.');
+        shell.exec('git branch -D gh-pages');
+    }
 }
